Remove problematic language
