+++ packages/aquarius/config.yml
@@ -3,7 +3,6 @@ owner: '90339695967350784'
 admins:
   - '82715425338560512'
   - '84108714671345664'
-blacklist: []
 home:
   guild: '91318657375825920'
   channel: '176846258218991616'
