+++ packages/aquarius/src/global/plugins/__test__/banlist.test.js
@@ -1,0 +1,5 @@
+test.todo('Checks banlist on start');
+
+test.todo('Checks banlist when joining a new guild');
+
+test.todo('Automatically leaves a guild that is included in the banlist');
