--- packages/aquarius/src/global/plugins/__test__/blacklist.test.js
@@ -1,5 +1,0 @@
-test.todo('Checks blacklist on start');
-
-test.todo('Checks blacklist when joining a new guild');
-
-test.todo('Automatically leaves a guild that is included in the blacklist');
